src/requirements.txt
annotated-types==0.6.0
ansible==12.0.0a4
ansible-core==2.19.0b4
anyio==4.3.0
bcrypt==4.0.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.7
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mongomock==4.3.0
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.18
packaging==24.0
passlib==1.7.4
pluggy==1.6.0
py4j==0.10.9.7
pyasn1==0.4.8
pycparser==2.21
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.6.3
pyspark==3.5.1
pytest==8.3.5
pytest-httpx==0.35.0
pytest-mock==3.14.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.19
pytz==2025.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
resolvelib==1.0.1
respx==0.22.0
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9
sentinels==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spark-nlp==5.3.2
starlette==0.40.0
typer==0.15.4
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
typing_extensions~=4.13.2
lxml~=5.4.0